Split register related declarations and impls
