+++ src/generate.rs
@@ -1,1 +1,2 @@
 pub mod generate;
+pub mod registers;
